move async tasks to structs, tail module
