+++ src/lib.rs
@@ -7,6 +7,7 @@ pub mod http;
 pub mod install;
 pub mod installer;
 pub mod settings;
+pub mod tail;
 pub mod terminal;
 pub mod upload;
 pub mod util;
